% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{calibration_out}
\alias{calibration_out}
\title{Generate model outputs for calibration from a parameter set}
\usage{
calibration_out(v_params_calib, l_params_all)
}
\arguments{
\item{v_params_calib}{Vector of parameters that need to be calibrated.}

\item{l_params_all}{List with all parameters of the decision model.}
}
\value{
A list with disease-free survival (DFS), disease-specific survival (DSS), and
overall survival (OS)
}
\description{
\code{calibration_out} computes model outputs to be used for calibration
routines.
}
