% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx2cea.R
\docType{package}
\name{darthpack}
\alias{darthpack}
\title{\code{darthpack} package}
\description{
The Decision Analysis in R for Technologies in Health (DARTH) coding framework
}
\details{
See the description of the package on
\href{https://darth-git.github.io/darthpack/index.html}{darthpack website}
or the README on
\href{https://github.com/DARTH-git/darthpack#readme}{GitHub}
}
