% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05b_deterministic_analysis_functions.R
\name{equal_breaks}
\alias{equal_breaks}
\title{Equal number of breaks}
\usage{
equal_breaks(n_breaks = 3, scale_factor = 0.05, ...)
}
\arguments{
\item{n_breaks}{Number of breaks}

\item{scale_factor}{Scaling factor}

\item{...}{Further arguments to function (not used)}

\item{parms}{Vector with strings with the name of the parameters of interest}

\item{ranges}{A named list of the form c("parm" = c(0, 1), ...) that gives
the ranges for the parameters of interest. The number of samples from this
range is determined by \code{nsamp}}

\item{nsamps}{number of parameter values. If NULL, 100 parameter values are
used}
}
\value{
A function that generates equal number of breaks across facets
}
\description{
This function runs a deterministic one-way sensitivity analysis (OWSA) on a
given function that produces outcomes. Obtained from:
https://stackoverflow.com/a/28459434
}
