% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{prior}
\alias{prior}
\title{Evaluate prior of calibrated parameters}
\usage{
prior(v_params)
}
\arguments{
\item{v_params}{Vector (or matrix) of model parameters}
}
\value{
A scalar (or vector) with prior values.
}
\description{
\code{prior} computes a prior value for one (or multiple) parameter set(s).
}
\examples{
v_param_names  <- c("r_DieMets",
                    "r_RecurCDX2pos",
                    "hr_RecurCDX2neg",
                    "p_Mets")
n_param        <- length(v_param_names)
v_lb <- c(r_DieMets       = 0.037, 
          r_RecurCDX2pos  = 0.001,
          hr_RecurCDX2neg = 1.58, 
          p_Mets          = 0.9))  # lower bound
v_ub <- c(r_DieMets       = -log(1-(1-0.03))/60, 
          r_RecurCDX2pos  = 0.03,
          hr_RecurCDX2neg = 4.72, 
          p_Mets          = 0.99) # upper bound
prior(v_params = sample.prior(n_samp = 5))
}
